+++ modules/dxp/apps/commerce/commerce-product-api/src/main/java/com/liferay/commerce/product/service/CommerceProductOptionValueLocalServiceUtil.java
@@ -57,7 +57,7 @@ public class CommerceProductOptionValueLocalServiceUtil {
 	public static com.liferay.commerce.product.model.CommerceProductOptionValue addCommerceProductOptionValue(
 		long commerceProductOptionId,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		long priority,
+		int priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService()
@@ -140,7 +140,7 @@ public class CommerceProductOptionValueLocalServiceUtil {
 	public static com.liferay.commerce.product.model.CommerceProductOptionValue updateCommerceProductOptionValue(
 		long commerceProductOptionValueId,
 		java.util.Map<java.util.Locale, java.lang.String> titleMap,
-		long priority,
+		int priority,
 		com.liferay.portal.kernel.service.ServiceContext serviceContext)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return getService()
